requirements.txt
langchain==0.1.0
langchain_community==0.0.12
langchain_openai==0.0.2
anthropic==0.8.1
chromadb==0.4.18
gitpython==3.1.40
python-dotenv==1.0.0
pydantic==2.5.2